Add meta tags to Collection HTML (#39357)
