+++ spec/requests/collections_spec.rb
@@ -6,12 +6,20 @@ RSpec.describe 'Collections' do
   describe 'GET /collections/:id' do
     subject { get collection_path(collection) }
 
-    let(:collection) { Fabricate(:collection) }
+    let(:collection) do
+      Fabricate(:collection, name: 'Frequent posters', description: 'Amazing people that can quickly fill your timeline', language: 'en')
+    end
 
-    it 'returns success' do
+    it 'returns success and includes opengraph meta tags' do
       subject
 
       expect(response).to have_http_status(200)
+
+      expect(head_meta_content('og:title')).to eq 'Frequent posters'
+      expect(head_meta_content('og:description')).to eq 'Amazing people that can quickly fill your timeline'
+      expect(head_meta_content('og:type')).to eq 'website'
+      expect(head_meta_content('og:url')).to eq collection_url(collection)
+      expect(head_meta_content('og:locale')).to eq 'en'
     end
   end
