+++ spec/requests/status_show_page_spec.rb
@@ -41,26 +41,5 @@ RSpec.describe 'Statuses' do
       expect(head_meta_content('og:locale')).to eq 'ca'
       expect(head_meta_content('og:description')).to eq status_text
     end
-
-    def head_link_icons
-      response
-        .parsed_body
-        .search('html head link[rel=icon]')
-    end
-
-    def head_meta_content(property)
-      response
-        .parsed_body
-        .search("html head meta[property='#{property}']")
-        .attr('content')
-        .text
-    end
-
-    def head_meta_exists(property)
-      !response
-        .parsed_body
-        .search("html head meta[property='#{property}']")
-        .empty?
-    end
   end
 end
